Use list_repo_files
