+++ vllm/transformers_utils/tokenizer.py
@@ -15,9 +15,12 @@ from typing_extensions import assert_never
 
 from vllm import envs
 from vllm.logger import init_logger
-from vllm.transformers_utils.config import get_sentence_transformer_tokenizer_config
+from vllm.transformers_utils.config import (
+    get_sentence_transformer_tokenizer_config,
+    list_filtered_repo_files,
+)
 from vllm.transformers_utils.tokenizers import MistralTokenizer
-from vllm.transformers_utils.utils import check_gguf_file, list_files_from_hf_or_path
+from vllm.transformers_utils.utils import check_gguf_file
 
 if TYPE_CHECKING:
     from vllm.config import ModelConfig
@@ -187,7 +190,7 @@ def get_tokenizer(
     mistral_common_installed = importlib.util.find_spec("mistral_common") is not None
     if tokenizer_mode == "auto" and mistral_common_installed:
         allow_patterns = ["tekken.json", "tokenizer.model.v*"]
-        files_list = list_files_from_hf_or_path(
+        files_list = list_filtered_repo_files(
             model_name_or_path=str(tokenizer_name),
             allow_patterns=allow_patterns,
             revision=revision,
