+++ vllm/transformers_utils/utils.py
@@ -1,7 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-import fnmatch
 import json
 import os
 import struct
@@ -9,8 +8,6 @@ from functools import cache
 from os import PathLike
 from pathlib import Path
 from typing import Any
-
-import huggingface_hub
 
 import vllm.envs as envs
 from vllm.logger import init_logger
@@ -132,67 +129,3 @@ def convert_model_repo_to_path(model_repo: str) -> str:
     from modelscope.utils.file_utils import get_model_cache_root
 
     return os.path.join(get_model_cache_root(), model_repo)
-
-
-def _list_local_hf_repo_files(repo_id: str, revision: str | None) -> list[str]:
-    """list the files of a local Hugging Face repo."""
-    repo_cache = Path(
-        huggingface_hub.constants.HF_HUB_CACHE
-    ) / huggingface_hub.constants.REPO_ID_SEPARATOR.join(
-        ["models", *repo_id.split("/")]
-    )
-
-    if revision is None:
-        revision_file = repo_cache / "refs" / huggingface_hub.constants.DEFAULT_REVISION
-        if revision_file.is_file():
-            with revision_file.open("r") as file:
-                revision = file.read().strip()
-
-    if revision:
-        revision_dir = repo_cache / "snapshots" / revision
-        if revision_dir.is_dir():
-            return [str(file) for file in revision_dir.glob("**/*") if file.is_file()]
-
-    return []
-
-
-def list_files_from_hf_or_path(
-    model_name_or_path: str,
-    allow_patterns: list[str],
-    revision: str | None = None,
-) -> list[str]:
-    """List files from Hugging Face Hub or local directory."""
-    assert len(allow_patterns) > 0
-    file_list = []
-
-    # Local dir
-    if os.path.isdir(model_name_or_path):
-        all_files = [
-            str(file)
-            for file in Path(model_name_or_path).glob("**/*")
-            if file.is_file()
-        ]
-
-    # Local HF repo
-    elif huggingface_hub.constants.HF_HUB_OFFLINE:
-        all_files = _list_local_hf_repo_files(
-            repo_id=model_name_or_path, revision=revision
-        )
-
-    # Remote HF repo
-    else:
-        hf_api = huggingface_hub.HfApi()
-        all_files = hf_api.list_repo_files(
-            repo_id=model_name_or_path, revision=revision
-        )
-
-    # Filter patterns on filenames
-    for pattern in allow_patterns:
-        file_list.extend(
-            [
-                file
-                for file in all_files
-                if fnmatch.fnmatch(os.path.basename(file), pattern)
-            ]
-        )
-    return file_list
